# swagger.json
{
  "swagger": "2.0",
  "info": {
      "version": "1.0.0",
      "title": "My Project CRUD",
      "description": "My Project Application API",
      "license": {
          "name": "MIT",
          "url": "https://opensource.org/licenses/MIT"
      }
  },
  "host": "localhost:8008",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/diagnosis/{id}": {
      "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of diagnosis",
            "type": "string"
          }
      ],
      "get": {
        "tags": ["Diagnosis by id"],
        "summary": "Get Diagnosis in system by id",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
  },
    "/diagnosis-delete/{id}": {
        "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "id of diagnostic",
              "type": "string"
            }
        ],
        "delete": {
          "tags": ["Delete diagnosis"],
          "summary": "Delete by id",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/diagnosis-create": {
      "parameters": [    
          {
            "name": "diagnosis",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "denumire_diagnostic": {
                  "type":"string"
                }
              }
            }
          }
      ],
      "put": {
        "tags": ["Create diagnosis"],
        "summary": "Create diagnosis",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      } 
      },
      "/diagnosis-update/{id}": {
        "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "id of diagnosis",
              "type": "string"
            },
            {
              "name": "med",
              "in": "body",
              "description": "the diagnosis to create",
              "schema": {
                "type": "object",
                "properties": {
                  "denumire_diagnostic": {
                    "type":"string"
                  }
                }
              }
            }
        ],
          "patch": {
            "tags": ["Update diagnosis by id"],
            "summary": "Update diagnosis in system by id",
            "responses": {
              "200": {
                "description": "OK"
              }
            }
          }
        },
...
          "/cities/{id}": {
            "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "description": "id of city",
                  "type": "string"
                }
            ],
            "get": {
              "tags": ["Cities by id"],
              "summary": "Get cities in system by id",
              "responses": {
                "200": {
                  "description": "OK"
                }
              }
            }
        },
          "/city-delete/{id}": {
              "parameters": [
                  {
                    "name": "id",
                    "in": "path",
                    "description": "id of city",
                    "type": "string"
                  }
              ],
              "delete": {
                "tags": ["Delete city"],
                "summary": "Delete by id",
                "responses": {
                  "200": {
                    "description": "OK"
                  }
                }
              }
            },
            "/city-create": {
            "parameters": [    
                {
                  "name": "city",
                  "in": "body",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id_judet": {
                        "type":"number"
                      },
                      "nume_oras": {
                        "type":"string"
                      }
                    }
                  }
                }
            ],
            "put": {
              "tags": ["Create city"],
              "summary": "Create city",
              "responses": {
                "200": {
                  "description": "OK"
                }
              }
            } 
            },
            "/city-update/{id}": {
              "parameters": [
                  {
                    "name": "id",
                    "in": "path",
                    "description": "id of city",
                    "type": "string"
                  },
                  {
                    "name": "city",
                    "in": "body",
                    "description": "the city to create",
                    "schema": {
                      "type": "object",
                      "properties": {
                        "id_judet": {
                          "type":"number"
                        },
                        "nume_oras": {
                          "type":"string"
                        }
                      }
                    }
                  }
              ],
                "patch": {
                  "tags": ["Update city by id"],
                  "summary": "Update city in system by id",
                  "responses": {
                    "200": {
                      "description": "OK"
                    }
                  }
                }
      },
...
      "/addresses/{id}": {
        "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "id of address",
              "type": "string"
            }
        ],
        "get": {
          "tags": ["Addresses by id"],
          "summary": "Get addresses in system by id",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
    },
      "/address-delete/{id}": {
          "parameters": [
              {
                "name": "id",
                "in": "path",
                "description": "id of address",
                "type": "string"
              }
          ],
          "delete": {
            "tags": ["Delete address"],
            "summary": "Delete by id",
            "responses": {
              "200": {
                "description": "OK"
              }
            }
          }
        },
        "/address-create": {
        "parameters": [    
            {
              "name": "address",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "id_oras": {
                    "type":"number"
                  },
                  "strada": {
                    "type":"string"
                  },
                  "cod_postal": {
                    "type":"string"
                  }
                }
              }
            }
        ],
        "put": {
          "tags": ["Create address"],
          "summary": "Create address",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        } 
        },
        "/address-update/{id}": {
          "parameters": [
              {
                "name": "id",
                "in": "path",
                "description": "id of address",
                "type": "string"
              },
              {
                "name": "address",
                "in": "body",
                "description": "the address to create",
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_oras": {
                      "type":"number"
                    },
                    "strada": {
                      "type":"string"
                    },
                    "cod_postal": {
                      "type":"string"
                    }
                  }
                }
              }
          ],
            "patch": {
              "tags": ["Update address by id"],
              "summary": "Update address in system by id",
              "responses": {
                "200": {
                  "description": "OK"
                }
              }
            }
   },
...
 }
}